Restructured model->mesh->submesh relationship.
