+++ game-engine/TODO.md
@@ -4,3 +4,4 @@
 - improve uniforms
 - IAssets may have to implement a getPath-method
 - make IGame or AGame to completely separate 'engine' package from the 'game' package
+- change name of model matrix (uModel) to object matrix (uObject) as model refers to a class that does not have a transform and thus no matrix of its own
