+++ game-engine/src/main/java/gameengine/game/Game.java
@@ -33,7 +33,8 @@ public class Game implements ITickable {
 	private void preloadAssets() {
 		AssetManager.Group preload = new AssetManager.Group("preload");
 		preload.put(new Texture("tex-default", FileUtils.getResourcePath("texture/texture.png")));
-		preload.put(new Model("mdl-default", FileUtils.getResourcePath("model/box.fbx")));
+		preload.put(new Mesh("mesh-man", FileUtils.getResourcePath("model/man.fbx")));
+		preload.put(new Mesh("mesh-box", FileUtils.getResourcePath("model/box.fbx")));
 		this.assetManager.registerGroup(preload);
 		preload.load();
 	}
