+++ game-engine/src/main/java/gameengine/util/ArrayUtils.java
@@ -3,6 +3,6 @@ package gameengine.util;
 public final class ArrayUtils {
 
 	public static boolean isInBounds(int index, Object[] array) {
-		return (index < 0 || index >= array.length);
+		return (index >= 0 && index <= array.length);
 	}
 }
